Export only minimal types for Pilets
