+++ src/packages/piral-cli/templates/default/pilet-index.tsx.ejs
@@ -1,7 +1,7 @@
-import { <%- apiName %> } from '<%- sourceName %>';
+import { PiletApi} from '<%- sourceName %>';
 import * as React from 'react';
 
-export function setup(app: <%- apiName %>) {
+export function setup(app: PiletApi) {
   app.showNotification('Hello from Piral!');
   app.registerMenu('sample-entry', () =>
     <a href="https://docs.piral.io" target="_blank">Documentation</a>
