+++ src/packages/piral-cli/templates/default/piral-index.jsx.ejs
@@ -2,15 +2,15 @@ import * as React from 'react';
 import { renderInstance, requestPiletsFromGraphQL, setupGqlClient } from 'piral';
 import { layout, errors } from './layout';
 
-const client = setupGqlClient();
+const gql = setupGqlClient();
 
 renderInstance({
   settings: {
-    gql: client,
+    gql,
   },
   layout,
   errors,
   requestPilets() {
-    return requestPiletsFromGraphQL(client);
+    return requestPiletsFromGraphQL(gql);
   },
 });
