+++ src/packages/piral-cli/templates/empty/pilet-index.tsx.ejs
@@ -1,4 +1,4 @@
-import { <%- apiName %> } from '<%- sourceName %>';
+import { PiletApi } from '<%- sourceName %>';
 import * as React from 'react';
 
-export function setup(app: <%- apiName %>) {}
+export function setup(app: PiletApi) {}
