+++ src/packages/piral-cli/templates/empty/piral-core-index.tsx.ejs
@@ -1,1 +1,1 @@
-export { PiletApi } from '<%- packageName %>';
+export * from '<%- packageName %>/api';
