+++ src/packages/piral-cli/templates/empty/piral-index.tsx.ejs
@@ -3,4 +3,4 @@ import { renderInstance } from 'piral';
 
 renderInstance();
 
-export { PiletApi } from '<%- packageName %>';
+export * from '<%- packageName %>/api';
