+++ src/packages/piral-containers/package.json
@@ -19,6 +19,7 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "files": [
+    "api.d.ts",
     "lib",
     "src"
   ],
