+++ src/packages/piral-core/api.d.ts
@@ -1,0 +1,6 @@
+export * from './lib/types/api';
+export * from './lib/types/components';
+export * from './lib/types/data';
+export * from './lib/types/extension';
+export * from './lib/types/meta';
+export * from './lib/types/utils';
