+++ src/packages/piral-ext/api.d.ts
@@ -1,0 +1,8 @@
+export * from 'piral-dashboard/api';
+export * from 'piral-feeds/api';
+export * from 'piral-fetch/api';
+export * from 'piral-menu/api';
+export * from 'piral-modals/api';
+export * from 'piral-notifications/api';
+export * from 'piral-translate/api';
+export * from 'piral-urql/api';
