+++ src/packages/piral-fetch/src/index.ts
@@ -1,3 +1,4 @@
+export * from './config';
 export * from './create';
 export * from './fetch';
 export * from './types';
