+++ src/packages/piral-fetch/src/types.ts
@@ -50,19 +50,6 @@ export interface FetchResponse<T> {
   text: string;
 }
 
-export interface FetchConfig {
-  /**
-   * Sets the default request init settings.
-   * @default {}
-   */
-  default?: RequestInit;
-  /**
-   * Sets the base URL to use for requests.
-   * @default location.origin
-   */
-  base?: string;
-}
-
 export interface PiletFetchApiFetch {
   /**
    * Triggers an actual HTTP/s request.
