+++ src/packages/piral-hyperapp/api.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/types';
