+++ src/packages/piral-menu/api.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/types';
