+++ src/packages/piral-menu/package.json
@@ -18,6 +18,7 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "files": [
+    "api.d.ts",
     "lib",
     "src"
   ],
