+++ src/packages/piral-ng/api.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/types';
