+++ src/packages/piral-search/api.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/types';
