+++ src/packages/piral-tracking/package.json
@@ -20,6 +20,7 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "files": [
+    "api.d.ts",
     "lib",
     "src"
   ],
