+++ src/packages/piral-urql/src/create.tsx
@@ -3,7 +3,36 @@ import { Extend } from 'piral-core';
 import { Client, Provider, defaultExchanges, subscriptionExchange } from 'urql';
 import { SubscriptionClient, OperationOptions } from 'subscriptions-transport-ws';
 import { gqlQuery, gqlMutation, gqlSubscription } from './queries';
-import { PiletGqlApi, GqlConfig, UrqlClient } from './types';
+import { PiletGqlApi, UrqlClient } from './types';
+
+export interface GqlConfig {
+  /**
+   * Sets the default request init settings.
+   */
+  default?: RequestInit;
+  /**
+   * Sets the URL of the GraphQL endpoint.
+   * @default location.origin
+   */
+  url?: string;
+  /**
+   * Sets the URL for the GraphQL subscription endpoint.
+   */
+  subscriptionUrl?: false | string;
+  /**
+   * Sets if the subscription should be lazy initialized.
+   */
+  lazy?: boolean;
+  /**
+   * Optional callback to the be used in case of a connection.
+   */
+  onConnected?(): void;
+  /**
+   * Optional callbsack to be used in case of a disconnect.
+   * @param err The connection error.
+   */
+  onDisconnected?(err: Array<Error>): void;
+}
 
 /**
  * Sets up an urql client by using the given config.
