+++ src/packages/piral-vue/api.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/types';
