+++ src/packages/piral/api.d.ts
@@ -1,0 +1,2 @@
+export * from 'piral-core/api';
+export * from 'piral-ext/api';
